feat: implement app cache deletion (#59)
* feat: implement app cache deletion

Introduce a `cache clean <app-id>` command to clean app cache.
This implementation also tries to stop the related app if running. In
case we do not care about that we can simply rm -rf `.cache/*`.

Closes #52

* feature: add `--force` flag

Add `--force` flag to forcefully terminate the app in case it's running,
and then clean the related cache.

* feat: add `clean-cache` under app

* feat: apply CR suggestion

* Update internal/orchestrator/cache.go

Co-authored-by: Luca Rinaldi <to@lucar.in>

* Update internal/orchestrator/cache.go

Co-authored-by: Luca Rinaldi <to@lucar.in>

---------

Co-authored-by: Luca Rinaldi <to@lucar.in>